ChafaCanvas: Remove obsolete code that was causing incorrect output.
There is no longer a need to threshold alpha, since ChafaPalette takes
care of it. Additionally, the extra palette lookup tended to make the
output worse, not better. Finally, it had a bug that caused very poor
output on edges next to alpha transparency with indexed colors and
high work factors.