ChafaCanvas: Fix issue with fill in FGBG mode
Solid symbols (U+2588) would be replaced by spaces because foreground
and background colors would be erroneously set to the same value,
indicating a "featureless cell" that should be replaced by blank_char
and the BG color set appropriately. However, in FGBG mode we cannot
change the background color, so don't try to do it there.