ChafaKittyCanvas: Specify target width/height in cells, version the seq
Kitty (version 0.19.2) behaves oddly when this is left unspecified and
the image dimensions are not an exact multiple of the cell size, as can
happen when you cat out a stored seq. This results in the cursor being
left in the top-left corner instead of to the right of the bottom-right
cell, which breaks inline animations that make use of relative cursor
movement.

Specifying the cell dimensions fixes the problem.

Additionally version the seq, since we may want to provide more
arguments in the future. Newer versions of Kitty have many interesting
features.