batch: Fix last row not being processed in some circumstances
Numerical imprecision could lead to miscalculating the index of the last
row, e.g. when rows/thread = 1.333333. This could cause the final row to
not be generated, and garbage would be emitted in its place.

The fix is to ensure the last batch always includes all remaining rows.

Fixes #87 (GitHub).