chafa: Proper workaround for ghostty emoji on echo off
Disable canonical mode along with echo. This works around the ghostty
heuristic that causes it to emit emoji when echo is turned off. The
relevant snippets from ghostty, for future reference:

.canonical = (attrs.c_lflag & c.ICANON) != 0,
.echo = (attrs.c_lflag & c.ECHO) != 0,

// We assume we're in some sort of password input if we're
// in canonical mode and not echoing. This is a heuristic.
const password_input = mode.canonical and !mode.echo;

Thanks to @veltza and @AnonymouX47 for discovering the issue and
suggesting the fix.

Part of fix for #239 (GitHub).