ChafaTerm: Don't set DISABLE_NEWLINE_AUTO_RETURN on win32
We want \n to act like \r\n when using g_print() to show help and
version information.

The official documentation is a little unclear between this flag
and ENABLE_WRAP_AT_EOL_OUTPUT, but it looks like this wasn't doing
anything that's useful to us in practice.

Fixes bug reported by @euamotubaina on the Matrix chat.