GridLayout: Properly handle finished chunks
Ensures `grid->finished_chunks` is set as soon as the grid runs out
of paths.

This fixes a bug where chafa runs infinitely (with high CPU usage)
when `--grid` is specified, a pixel graphics output format is used,
and all input paths are invalid (do not exist, of unknown format,
unable to load, etc).