chafa: Remove ConHost support
This experimental mode would draw directly to character cells in
the legacy ConHost mode on MS Windows. However, it didn't support
most of the command-line options, and acted directly on a drawing
surface instead of producing a data stream, which made it a bit
of an awkward fit.

The original author has declined to maintain it, and since no one
else stepped up, I'm removing it after the one-month discussion
window closed.

Fixes #287 (GitHub).